+++ src/roots.rs
@@ -1,9 +1,6 @@
 use std::path::{Path, PathBuf};
 use std::{env, fs};
 
-use dialoguer::theme::ColorfulTheme;
-use dialoguer::Select;
-
 use crate::database::Database;
 use crate::error::FsPulseError;
 use crate::schedules::{delete_schedules_for_root_immediate, root_has_active_scan_immediate};
@@ -19,32 +16,6 @@ pub struct Root {
 }
 
 impl Root {
-    pub fn interact_choose_root(prompt: &str) -> Result<Option<Root>, FsPulseError> {
-        let conn = Database::get_connection()?;
-        let mut roots = Root::roots_as_vec(&conn)?;
-        if roots.is_empty() {
-            print!("No roots in database");
-            return Ok(None);
-        }
-
-        let mut labels: Vec<&str> = roots.iter().map(|root| root.root_path()).collect();
-        labels.push("Exit");
-
-        let selection = Select::with_theme(&ColorfulTheme::default())
-            .with_prompt(prompt)
-            .default(0)
-            .items(&labels)
-            .interact()
-            .unwrap();
-
-        // "Exit" is the last option in the prompt
-        if selection == roots.len() {
-            Ok(None)
-        } else {
-            Ok(Some(roots.remove(selection)))
-        }
-    }
-
     pub fn get_by_id(conn: &Connection, root_id: i64) -> Result<Option<Self>, FsPulseError> {
         conn.query_row(
             "SELECT root_path FROM roots WHERE root_id = ?",
@@ -86,42 +57,6 @@ impl Root {
 
     pub fn root_path(&self) -> &str {
         &self.root_path
-    }
-
-    pub fn roots_as_vec(conn: &Connection) -> Result<Vec<Root>, FsPulseError> {
-        let mut roots: Vec<Root> = Vec::new();
-
-        Root::for_each_root(conn, |root| {
-            roots.push(root.clone());
-            Ok(())
-        })?;
-
-        Ok(roots)
-    }
-
-    pub fn for_each_root<F>(conn: &Connection, mut func: F) -> Result<(), FsPulseError>
-    where
-        F: FnMut(&Root) -> Result<(), FsPulseError>,
-    {
-        let mut stmt = conn.prepare(
-            "SELECT root_id, root_path
-            FROM roots
-            ORDER BY root_id ASC",
-        )?;
-
-        let rows = stmt.query_map([], |row| {
-            Ok(Root {
-                root_id: row.get::<_, i64>(0)?,
-                root_path: row.get::<_, String>(1)?,
-            })
-        })?;
-
-        for row in rows {
-            let root = row?;
-            func(&root)?;
-        }
-
-        Ok(())
     }
 
     /// Delete a root and all associated data (scans, items, changes, alerts, schedules).